requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.6.0
django-crequest==2018.5.11
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2==2.9.10
PyJWT==2.10.1
python-decouple==3.8
setuptools==75.1.0
sqlparse==0.5.2
tzdata==2024.2
wheel==0.44.0
